Introduce --output option
Prior to this change, any files attached to a message are printed to the
standard output as
```
    name: private_url
```
It is then the user's choice to decide how to access the file at the
provided URL.

This change introduces a --output (-o) option that allows the user to
specify a file to download and write all files or named files.  By
default nothing is done and the behaviour is the same when the option is
not specified.  Otherwise, the string passed to the option is
interpreted as name-value pairs separated by a colon (:), where the name
is the same as reported in slack the slack message and the value is a
file to write the attachment to.  If a bare value is specified then it
will match all attachments.

Signed-off-by: Corin Lawson <corin@responsight.com>